Move module creation code to a module factory
All modules are going to be builtin for now as it's simpler than
defining a plugin mechanism at the moment. In the future, there will be
a plugin mechanism, but basic stuff is going to be builtin for
performance reasons.

The module creation functions will be registered in the module factory
and created through that instead of through a hard-coded switch
statement in the loader.