Begin work on the vanguard inventory
This is an idea directly copied from ansible. I think it's a good idea
so we're going to do it too. We want to be a remote cluster management
tool too. Just because we're going to have a local agent on the machine
doesn't mean we shouldn't have a good remote client, and the remote
client is going to need to know how to discover multiple machines.